refactor: consolidate llm provenance helpers and fix naming inconsistencies
Per-thread pipeline refactor-clean pass on post-v2.0.0 code.

MF-1: Move validateIntentionallyBlank from extract.go to shared.go — used
by all three parse functions; wrong home was a maintenance risk.

MF-2/MF-4: Extract stampProvenance helper in shared.go, replacing the
identical 10-line provenance-stamping block that was duplicated verbatim
across extractSingleDoc, parseSingleDraft, and parseCritiqueDraft. Caller
supplies ExtractionStage ('weak-draft' or '') to preserve the intentional
divergence. Eliminates the goto stamp pattern in parseCritiqueDraft (MF-4),
which now uses the same boolean-sentinel pattern as parseSingleDraft.

MF-3: Rename stringSlicesEqual → stringSlicesEqualOrdered (draftchain.go)
and slicesEqual → stringSlicesEqualUnordered (extractiongap.go). The old
names gave no signal that one is order-sensitive and the other is not.
Cross-reference comments added to each.

MF-5: Add explicit int64() cast to LimitReader in prompt.go, matching the
existing pattern in shared.go. No behaviour change; improves audit clarity.

MF-6/NH-6: Move splitErrNotes/joinErrNotes from assist.go to shared.go.
Use joinErrNotes in critique.go instead of strings.Join inline.

NH-3: Add +1 to LimitReader in readSessionFile and add the corresponding
oversize check, matching the pattern in readSpansFile. Prevents silent
truncation of exactly-at-limit session files.

+++ meshant/llm/prompt.go
@@ -16,7 +16,7 @@ func LoadPromptTemplate(path string) (string, error) {
 	}
 	defer f.Close()
 
-	limited := io.LimitReader(f, maxSourceBytes+1)
+	limited := io.LimitReader(f, int64(maxSourceBytes)+1) // +1 to detect oversized files
 	data, err := io.ReadAll(limited)
 	if err != nil {
 		return "", fmt.Errorf("llm: read prompt template %q: %w", path, err)
